requirements.txt
numpy>=1.15.4
opencv-python>=3.4.3.18
chainer>=5.0.0
cupy>=5.0.0
chainerrl>=0.5.0
future~=0.18.2
Pillow~=9.1.1
scipy~=1.7.3
matplotlib~=3.5.2
torch~=1.12.0
torchvision~=0.13.0
scikit-image~=0.19.3
sewar~=0.4.5